(MAINT) Add PR Checklist GHA (#5)
* (MAINT) Add PR Checklist GHA

This change adds the GHA for verifying contributor PRs against the
PR Template checklist, failing a PR if the checklist is not filled
out fully.

It also converts previously CRLF files to LF to prevent future
complaining from git.

* (MAINT) Rename PSRule actions for clarity

* (MAINT) Add copyright/license headers to code files

* (MAINT) Add support and contributing stubs

* (MAINT) Default to LF for file creation

This commit adds a VS Code setting to default to creating files in this
project with an LF line ending instead of using the OS default. This
standardizes editor behavior with git attribute configuration.