(MAINT) Reorganize source files into shared folder
Prior to this change, the source files for the PowerShell modules,
cSpell dictionaries, and Vale styles lived in their own folders at the
top level. This change colocates them in a new `Source` folder to limit
the number of top-level files and help with organization.

This paves the way for adding the documentation and the docs site files
to the project.