(MAINT) Respect EOL, notices, & path in preamble
Prior to this change, the `WriteMungedContentWithReferencePreamble()`
method in the **SourceReference** class didn't respect the file's
original EOL characters or included copyright and license notices.

That meant that when using the `Update-SourceDependency` cmdlet that the
updated files would lose their notices and have their EOL munged on
Windows if it wasn't already set to CRLF.

Further, the logic for writing the required functions in the preamble
was implemented to separate the path on the _first_ folder named
`Source`, not the last. This previously worked when the module was
defined at `Modules/Documentarian.DevX/Source` but broken when the
project was reorganized to place all code in the top-level `Source`
folder.

This change:

1. Updates the `WriteMungedContentWithReferencePreamble()` method to
   use the new **LineEnding**, **CopyrightNotices**, and
   **LicenseNotices** properties defined on a **SourceFile**. It ensures
   that the EOL for the newly munged content is the same as the original
   file and that the notices are included before any other statements.
2. Updates the `ResolveReferencePreamble()` method to find the source
   folder for the processed file and build the correct path instead of
   the previous naive regex implementation.
3. Ensures that using the `Update-SourceDependency` cmdlet correctly
   updates dependencies without the previous side effects of stripping
   the copyright and license notices, munging the file's EOL, and
   writing the wrong path for required functions.