(CI) Add workflow to build modules on PR/Merge
Prior to this change, all building of the PowerShell modules was done
locally for verification and prototype validation.

This change adds a new workflow to the repository for building the
PowerShell modules and archiving them as artifacts on PR and merge to
main so they can be tested and inspected without a local build.

This is a prerequisite step to being able to publish the modules.