(MAINT) Add vale sync task
Prior to this change, using our own definitions for Vale
required a dance to pull the correct styles when syncing
and required vendoring external styles.

This change takes advantage of the latest release of Vale
to enable installing a style from the local folder and
defines both an Invoke-Build task and a VSCode task for
synchronizing the vale styles.