(MAINT) Add docs/build for vale styles
This change overhauls the project to account for documenting and
publishing vale style packages. It updates the project tasks and
build scripts to support the changes.

It adds two new shortcodes to the site for documenting vale rules and
(re)publishing the package to the site.

It also overhauls the project build scripts and adds support for using
the same build scripts locally, in Netlify, and in GitHub Actions.