(MAINT) Fix Vale publishing
Prior to this change, the published Vale package didn't actually work
when specified as a package in a Vale configuration file.

The structure for 'complete' Vale packages is very specific and our
publishing did not adhere to the structure. For a Vale package to work:

1. The zip file must contain a folder with the same name as the zip and
   style package, like `PowerShell-Docs`.
1. The `.vale.ini` must be in that folder.
1. The **StylesPath** value for the `.vale.ini` file must be `styles`
1. The rules must be defined as `.yml` files in a folder with the same
   name as the style package _inside_ the `styles` folder.

For example, this structure is valid:

```tree
PowerShell-Docs.zip/
├─ PowerShell-Docs/
│  ├─ .vale.ini
│  ├─ styles/
│     ├─ PowerShell-Docs/
│       ├─ Avoid.yml
```

To support publishing a usable Vale style package, this change:

1. Moves the definitions for the **PowerShell-Docs** style package into
   the `Source` folder, making it easier to understand exactly what is
   used to define the package.
1. Updates the build script to handle correctly building the package
   for publishing it.
1. Updates the site configuration to reflect the updated structure.