(MAINT) Rename Source folder to Projects
This change renames the `Source` folder in the repo to `Projects`
to make it easier to navigate the repository and understand when
you are looking at source files vs overall project files.

This rename is more semantically accurate, as these folders contain
the project source files as well as tests and documentation.