(GH-71) Strongly type output for `Get-Vale`
Prior to this change, the `Get-Vale` command returned an **ApplicationInfo**
object with the version set to `0.0.0.0` regardless of the actual version.

This change adds a new type to the module, **ValeApplicationInfo**, which
has the same properties as **ApplicationInfo**. It does not inherit from
**ApplicationInfo** due to an inability to workaround the constructor
failing.

Instead, it sets the properties to the same as defined on the **ApplicationInfo**
object and overwrites the value of the **Version** property to match the output
from running `vale --version`.