(MAINT) Fix using statement in ModuleComposer
Prior to this change, the `ComposeInitScriptContent()` method
in the **ModuleComposer** class of the **DevX** module used a
forward slash for the path to the module, as would the `using`
statement added for the private module in the
`GetUsingPrivateModuleStatement()` method.

In older versions of PowerShell, the `using` statement requires
backslashes for relative paths. We confirmed that using a `\`
still works on non-Windows platforms.

This enables the modules to load correctly in older versions of
PowerShell.