(MAINT) Add docs composer to DevX
Prior to this change, the DevX module had no special handling
for documentation in composed modules.

This change adds the new `ComposeDocumentation()` method with
a naive implementation to at least compose the cmdlet reference
documentation. It has no special handling and ignores about
topics, which will need to be addressed in the future.

For now, this should serve as a first step to ensuring the
modules ship with documentation and can be used externally.