(MAINT) Rework Vale docs to comply with PlatyPS
Prior to this change, the cmdlet reference documentation for
the Documentarian.Vale module was unexportable with PlatyPS.
The documents used Hugo shortcodes, like `xref`, were missing
an explicit H1, had a newline after the Synopsis header, and
used reference links.

This change munges the content to conform with PlatyPS. In the
future, it would be preferable to use better formatting and the
advanced syntax available in Hugo without breaking the console
documentation. For now, shipping the docs is more important.