(MAINT) Strongly type Readability results
Prior to this change, the output for `Get-Readability` was
a **PSCustomObject**. This change reworks the cmdlet and
introduces new types for a more robust output.

- **ValeReadabilityRule** is a new enum that defines the
  known and implemented readability rules, borrowing from
  Vale and their official definition.
- **ValeReadability** is a new base class with useful static
  members for working with readability metrics.
- The `ValeMetrics*` classes inherit from **ValeReadability**
  to add specific types for each of the readability rules.

  These classes handle their own scoring, defaults, and have
  dynamic members where sensible.

This change also introduces a new pattern of using a dummy
property on the class definition and overwriting it in the
constructor with a **ScriptProperty**. This enables us to
have dynamic properties as well as custom getter/setters.

This change also takes advantage of type decoration to provide
multiple useful views over the same objects without the user
needing to specify which view they want.