(MAINT) Improve error handling for `Invoke-Vale`
Prior to this change, the `Invoke-Vale` cmdlet did not handle the error
case where no `.vale.ini` file was found and the `--config` flag was not
used. Several error messages also reported only the error code instead of
the actual error message from Vale. Finally, when the experimental
`$PSNativeCommandErrorAction` preference variable was set to `$true`, the
error handlers would not be invoked correctly.

This change adds a new handler for missing default config and ensures the
other handlers report a message instead of the code. It temporarily
sets the `$PSNativeCommandErrorAction` to enable our error-handling
across versions.

Finally, it updates the **Parameter** attribute to include
`ValueFromRemainingArguments` to enable the intended behavior.