(MAINT) Refactor `Get-ProseMetric`
This change refactors the `Get-ProseMetric` cmdlet for readability
and UX. It slightly increases performance by ensuring that the
processing is all streamed instead of collected-and-then-looped.

It replaces the potentially surprising behavior (automatic recursion)
with an explicit **Recurse** parameter and adds the correct attributes
to the **Path** parameter.

It also fixes a bug in the formatter so the File name actually displays.