(MAINT) Update `Get-Readability` to `Get-ProseReadability`
This change renames the `Get-Readability` cmdlet to `Get-ProseReadability`
and refactors it for simiplicity and to align with the updated implementation
of `Get-ProseMetric`.

Instead of having grouped and ungrouped formats, it now always groups the
results by folder. The new **Recurse** parameter reflects the same parameter
in `Get-ProseMetric`.

Finally, this change fixes a bug in the processing to ensure that non-Markdown
files do not end the overall processing but instead move to the next object in
the pipeline as intended.