(GH-91) Add `Install-Vale` command
Prior to this change, the **Documentarian.Vale** module included the
`Install-WorkspaceVale` command, which enabled users to install Vale
into their current working directory.

However, this required users to install Vale to every project, which
also meant needing to modify the `.gitignore` file or continually
manage an untracked folder in the project root.

This change:

- Renames `Install-WorkspaceVale` to `Install-Vale`.
- Adds support to `Get-Vale` for finding an installation in the user's
  home directory. If Vale isn't found in the workspace,  the command now
  searches for `~/.vale/vale` before searching the `PATH`.
- Modifies the logic for searching in the workspace scope to recursively
  search up the file tree from the current working directory instead of
  only searching the current working directory.
- Adds the `ValeInstallScope` enumeration, which is used by
  `Install-Vale` to determine where to install Vale. By default, the
  command installs Vale to the `Workspace` scope.
- Updates the documentation.