(MAINT) Add changelog link shortcodes
This change adds new shortcodes to the docsy-pwsh Hugo module for
easier changelog link writing.

The `changelog/link/source` shortcode adds a link to the source
code for a given tagged release, using the repository, tag, and
source path.

The `changelog/link/prs` shortcode adds a link to the pull requests
for a given release, building the pull request search query so the
source does not need to include the unreadable URL queries.