(MAINT) Scaffold MarkdownBuilder module
This change introduces another module to the project. In several of the
other modules, we found a need for building Markdown strings programmaticaly
in our functions.

The new module, **Documentarian.MarkdownBuilder**, adds a new class and several
helper functions to make using **System.Text.StringBuilder** to write
arbitrary Markdown easier.

The functions are designed to be chainable in a pipeline without requiring the
author to carefully control the output. By default, the functions (except for
`New-Builder`) never return any output. When they are called in a pipeline and
are not the last command in the pipeline, they automatically emit the
**MarkdownBuilder** they are using. Authors can specify the **PassThru**
parameter to force the functions to output the **MarkdownBuilder**.

The **MarkdownBuilder** has built-in handling for ensuring that code blocks are
valid Markdown, even when they are nested. The helper classes **LineEnding** and
**CodeFenceCharacter** provide friendly ways to specify values for their
associated enums and transform the enum values into the correct string.

In the future, other Documentarian modules should be able to take a dependency
on **Documentarian.MarkdownBuilder** for writing Markdown strings in their
functions.

This module is separate from the usage of **Documentarian**, which parses and
modifies Markdown documents, because its about writing the _content_, often
before any document exists or dynamically for a specific purpose.