(MAINT) Scaffold AstInfo module
This change scaffolds the new **Documentarian.AstInfo** module. This
module provides a more robust and flexible way to investigate the AST
of PowerShell code.

We were using prototypes of the classes and functions in this new
module in the **DevX** module. However, we found a need for this
functionality in other modules. In the future, **DevX** should be
refactored to use this new module.

That also limits the concerns of the DevX module to how it handles
source code, not the parsing of the AST itself.

With this new module and the **MarkdownBuilder** module, we are able to
begin the work of adding the classes and functions to **ModuleAuthor**
for parsing PowerShell source code to retrieve documentation info and
approach the problem of documenting classes, enums, DSC Resources, and
other structured definitions in PowerShell.