(MAINT) Update DevX template for build script
This change updates the template for the DevX build script laid down
in the `New-ComposableModule` function. It also brings the newly
scaffolded modules in line with the build scripts for the other
modules.

The updated script imports the bundled tasks from the DevX module
instead of copying a version of those tasks into the build script.