(MAINT) Fix AstInfo to support non-scriptblocks
Prior to this change, the **AstInfo** class expected the AST to
always be a **ScriptBlockAst** - convenient when parsing a file,
but limits the reusability of the wrapper object. This change
sets the type to **Ast** instead, which **ScriptBlockAst**
inherits from.

This change ensures passing a non-scriptblock AST to the class
constructors for text and arbitrary AST objects does not throw
an error and works as expected.