(MAINT) Remove AstInfo
In the process of this work, it made more sense to colocate the AstInfo
functionality, the DecoratingComments functionality, and the HelpInfo
functionality in a single module. AstInfo is made much more functional
by including the DecoratingComments, and HelpInfo depends almost
entirely on the DecoratingComments and AstInfo.

This change remmoves the functionality from the unreleased module in
preparation for adding them to ModuleAuthor.