(MAINT) Reimplement module type autoloading in DevX
This change reimplements the approach to ensuring users can call
the public classes and enums defined in a module automatically
after import without needing a separate using statement.

Instead of placing a using statement in the init script, the
new implementation creates TypeAccelerators that map to the
public types in the module. This is even more effective than
before, because it does not clobber the module types _and_
immediately provides intellisense, not just functionality.