(MAINT) Implement HelpInfo, AstInfo, and DecoratingComments
This change adds a suite of new functions and classes to ModuleAuthor, enabling
developers to inspect their PowerShell code and prototyping a new method for
documenting their code. The design will evolve in the near future.

In this implementation, you can document and retrieve documentation for your
PowerShell classes and enums. We do not yet have a formatter for the output
data to make use of it. That will follow in a future change.