(MAINT) Add basic acceptance and fix bugs
This change adds an initial set of basic acceptance tests for the new functionality,
validating the documentation behavior for classes. By testing the resolution of a
class without comments, with non-parsing comments only, and with a complex mixture
of parsing and non-parsing comments, the tests help us validate the AstInfo,
DecoratingComments, and HelpInfo functionality at a high-level.

In the future, we should expand the test cases. In particular, we need to cover the
Enum resolution.

This change still does not include documentation, which will block its release.
However, this change does make the functionality "complete" so it can be merged and
iterated on.