(MAINT) Update decorating comments
This change reimplements the finding logic for decorating comments
to work on tokens and prefer AST objects or AstInfo objects as input,
adding additional handling to avoid edge cases for finding comments,
like:

- When a comment is added beside a property or label and immediately
  above the decorating comment for the next property or label.
- Handling properties and labels with a decorating comment beside them.
- Only using internal block comments when they are multi-line.
- Correctly trimming one-line block comments
- Fixing a bug in joining multi-line non-block comments

This change also makes minor fixes to Enum help, ensuring that the docs
for a Label are preferred over direct comment decoration and bringing
the resolve logic more in line with other implementations.

Finally, this change adds acceptance tests for generating help from the
decorating comments in an Enum definition.