(GH-115) Replace powershell-yaml dep with YaYaml
Prior to this change, the **Documentarian** module had an explicit
dependency on the **powershell-yaml** module and **Documentarian.Vale**
had an implicit dependency.

This dependency caused a conflict with **PlatyPS**, which also depends
on the YamlDotNet library that **powershell-yaml** uses. This conflict
prevented the modules from working correctly for normal workflows like
exporting and updating documentation. It also meant that anyone loading
either of these modules was prevented from loading PlatyPS in the same
session.

This change replaces the dependency on **powershell-yaml** with the
**YaYaml** module instead. **YaYaml** uses an assembly load context
to load the YamlDotNet library, preventing the conflict.

This change should make the modules more functional, no longer
requiring users to operate in two different sessions to manage their
PowerShell documentation.