(GH-130) Improve type accelerator handling
Prior to this change, the DevX module would create type accelerators
for every public class and enum in the module, making them available
on module load without the need for an initialization script or for
the user to call `using module`. However, the generated code didn't
handle unregistering the type accelerators when a user removes the
module. The composition process also always injects the initialize
script to run, even when it doesn't have any statements and is just
the default template.

This change:

- Overhauls the type accelerator handling, ensuring that the build
  generates code for the root module to list the public types,
  check if a type with the same name is already registered as a
  type accelerator (throwing an error if so), register them as
  type accelerators on module load, and remove them with the module.
- Updates the logic in composition to check if the initialization
  script has any code, ignoring it if it doesn't.
- Resolves #130