(MAINT) Add HelpInfoFormatter to ModuleAuthor
This change adds the **HelpInfoFormatter** class, which is used for creating
default and custom formatters that transform instances of the `*HelpInfo`
classes into output documentation.

This change also adds the **HelpInfoFormatterDictionary** class as a special
dictionary that the `*HelpInfo` classes can use to define and instantiate
their own built-in formatters. Users can then leverage, extend, or replace
the formatters to suit their needs.